+++ packages/terra-action-header/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to be compatible with `react-intl` v2-v5
+  * Removed usage of `FormattedMessage` in favor of `injectIntl`
+
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 2.61.0 - (December 8, 2020)
 
 * Changed