+++ packages/terra-icon/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Changed
   * Bumped `one-cerner-style-icons` to v1.35.0
+  * Updated to be compatible with `react-intl` v2-v5
 
 ## 3.38.0 - (November 17, 2020)