+++ packages/terra-show-hide/CHANGELOG.md
@@ -8,6 +8,8 @@
 ## 2.46.0 - (December 8, 2020)
 
 * Changed
+  * Updated to be compatible with `react-intl` v2-v5
+  * Removed usage of `intlShape`
   * Update wdio tests to remove deprecated Terra.it helpers
 
 ## 2.45.0 - (November 17, 2020)